Extract EmailPayloadTrait from AbstractEmails
Move getPayload() and its address/attachment/header helpers into a
reusable trait so the inbound reply/forward API can map a Symfony Mime
Email to the Mailtrap payload the same way the send API does. Pure
refactor — no behavior change (send/batch tests unchanged).
